requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
asn1crypto==0.24.0
bsdiff4==1.1.5
cffi==1.11.5
coverage==4.5.2
cryptography==2.4.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-nose==1.4.6
django==2.1.4
djangorestframework==3.9.0
idna==2.8
nose==1.3.7
pycparser==2.19
pymysql==0.9.2
pytz==2018.7
six==1.12.0
